+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [2.2.1] - 10-10-23
+
+- Added _SetTextureOverride, _GetTextureOverride, _GetResolvedTextureOverride to ScreenManager
+- Set minimum CommonTXL version to 1.0.3 (hopefully get around VCC version resolution issues)
+
 ## [2.2.0] - 10-04-23
 
 - Added support for multiple CRT definitions